fllood-predict/requirements.txt
matplotlib==3.4.3
numpy==1.20.3
pandas==1.4.3
Pillow==9.2.0
seaborn==0.11.2
streamlit==1.12.0
sklearn